refacor country
